feat(tools): move processing to a dedicated dir
