+++ geoplateforme/processing/tools/sld_downgrade.py
